requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database drivers
motor>=3.3.2
pymongo>=4.6.0

# Vector database
qdrant-client==1.9.1

# Pydantic for data validation
pydantic>=2.8.0
pydantic-settings>=2.4.0

# HTTP client for external API calls
httpx>=0.25.2

# Environment and configuration
python-dotenv>=1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.13.0
flake8==6.1.0

# LLM Providers
openai>=1.35.0
google-generativeai>=0.7.2

# File uploads
python-multipart>=0.0.6

# Audio processing and speech recognition
speechbrain>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
ffmpeg-python>=0.2.0
pydub>=0.25.0
resemblyzer>=0.1.1
audioread>=3.0.0
resampy>=0.4.0
numba>=0.56.0